feat(capsule): carry authenticating device key_id from handshake to publish-as
Extend the per-connection identity mechanism that derives the verified
principal (issue #45/#852) to also carry the matched device key_id, so a
paired device's scope can be enforced at the kernel cap-gate.

- handshake: verify_signature_against_keys returns the matched DeviceKey's
  key_id; propagated through verify_principal_signature, run_principal_challenge,
  and validate_handshake (now yields (PrincipalId, key_id)).
- registry: ConnectionIdentity { principal, device_key_id } replaces the
  bare PrincipalId value so the two never desync; accept/poll-accept bind both.
- tcp-stream.read: set/clear ingress_principal AND ingress_device_key_id in
  lockstep on data/non-data reads, so a stale id never leaks onto a later forward.
- publish-as stamps the host-derived device_key_id onto IpcMessage; publish
  (own-principal) never stamps one.
- split the ipc host test module into ipc_tests.rs to stay under the line cap.
