+++ crates/astrid-capsule/src/engine/wasm/test_fixtures.rs
@@ -140,6 +140,7 @@ pub(crate) fn minimal_host_state(rt: tokio::runtime::Handle) -> HostState {
         connection_principals: Arc::new(dashmap::DashMap::new()),
         client_connections: Arc::new(dashmap::DashMap::new()),
         ingress_principal: None,
+        ingress_device_key_id: None,
         recv_yielded: false,
         no_yield_windows: 0,
     }
